Support for setting a default password.
It is now possible to set a default password during configuration. This is
useful when console access is not always possible and remote network access
requires password authentication. For example, the Dropbear SSH server
refuses to login use 'root' if this user has no password set.

The DHCP client now also sends the hostname to the DHCP server.

Some sample configurations are included. The 'minimal' configuration
consists of a default Busybox and Buildroot configuration. The 'ssh'
configuration enables the Dropbear SSH server and sets the default password
for the 'root' user to 'bifferboard'.

Use the provided sample configuration by running:

  $ make bifferboard_ssh_defconfig
  $ make